Massive Redux refactor
I know, I know, this is a massive patch that does a shitload of things
all at once. Deal with it.

The main list view is now using Redux to toggle the state of
individual escalators. Yet to be done is to build a container around
the status output so that it can plug into the store and self-update
as well.