Inject history into escalator state data
We don't return all escalator history with each escalator's main data,
but we do store it within the escalator's item in the shared
state. This may or may not be the "right" way to do it, but it works.

At first, the items will not have a "history" key at all, so in that
case, create it and then populate the appropriate sub-key with the
data retrieved.